LPS-54098 Move BaseLocalServiceTreeTestCase to its original package
