+++ portal-impl/test/integration/com/liferay/portal/service/GroupLocalServiceTreeTest.java
@@ -20,7 +20,7 @@ import com.liferay.portal.kernel.test.util.TestPropsValues;
 import com.liferay.portal.model.Group;
 import com.liferay.portal.model.GroupConstants;
 import com.liferay.portal.model.TreeModel;
-import com.liferay.portal.service.tree.test.BaseLocalServiceTreeTestCase;
+import com.liferay.portal.service.test.BaseLocalServiceTreeTestCase;
 import com.liferay.portal.test.rule.LiferayIntegrationTestRule;
 import com.liferay.portal.test.rule.MainServletTestRule;
